requirements.txt
Django
Pillow
asgiref
pip
pytz
setuptools
sqlparse
django-bootstrap4
pandas
django-simple-captcha











